Separate TabView types into different files
